Fix define-pyfun macro in compiled mode
The macro would expand to a direct reference to the
hash-table-ref/default and hash-table-set! procedures rather than the
identifiers for those procedures.  This works in the interpreter
because it is a bit lax when intermixing expansion code with evaluated
code, but when compiling code that uses the macro, it breaks with this
error:

  Syntax error: illegal atomic form

	#<procedure>

Fixes issue #12